+++ Sources/StreamChat/StateLayer/ChannelListState.swift
@@ -7,7 +7,11 @@ import Foundation
 
 /// Represents a list of channels matching to the specified query.
 @MainActor public final class ChannelListState: ObservableObject {
-    private let observer: Observer
+    private var observer: Observer
+    private var shouldSkipInitialRemoteUpdate = false
+    private var handlers: Observer.Handlers {
+        .init(channelsDidChange: { [weak self] in self?.channels = $0 })
+    }
     
     init(
         query: ChannelListQuery,
@@ -28,14 +32,29 @@ import Foundation
             eventNotificationCenter: eventNotificationCenter,
             channelWatcherHandler: channelWatcherHandler
         )
-        channels = observer.start(
-            with: .init(channelsDidChange: { [weak self] in self?.channels = $0 })
-        )
+        channels = observer.start(with: handlers)
     }
     
     /// The query used for filtering the list of channels.
-    public let query: ChannelListQuery
+    public private(set) var query: ChannelListQuery
     
     /// An array of channels for the specified ``ChannelListQuery``.
     @Published public internal(set) var channels: [ChatChannel] = []
+
+    func skipNextInitialRemoteUpdate() {
+        shouldSkipInitialRemoteUpdate = true
+    }
+
+    func consumeShouldSkipInitialRemoteUpdate() -> Bool {
+        defer { shouldSkipInitialRemoteUpdate = false }
+        return shouldSkipInitialRemoteUpdate
+    }
+
+    func reset(
+        query: ChannelListQuery,
+        minimumFetchLimit: Int
+    ) {
+        self.query = query
+        channels = observer.resetFetchRequest(query: query, minimumFetchLimit: minimumFetchLimit)
+    }
 }